requirements.txt
fastapi
uvicorn
websockets
pytest
PyYAML
requests
websocket-client
psutil
discord.py
python-telegram-bot
